requirements.txt
Django==5.2.7 ,
djangorestframework==3.16.1,
asgiref==3.10.0,
sqlparse==0.5.3,
pytz==2022.1,
mysql-connector-python==9.1.0,
drf-yasg==1.21.5,
mysqlclient==2.2.7,
django-credit-cards==0.4.1